Reactor MFE cas220101 into an external costing function
