+++ pyfecons/costing/calculations/cas22/cas220101_reactor_equipment.py
@@ -3,13 +3,66 @@ from io import BytesIO
 from typing import Dict
 from matplotlib import pyplot as plt
 from pyfecons.costing.calculations.conversions import to_m_usd
+from pyfecons.costing.calculations.volume import calc_volume_ring, calc_volume_torus
 from pyfecons.data import CAS220101
 from pyfecons.enums import ReactorType, BlanketFirstWall, BlanketType
 from pyfecons.inputs import RadialBuild, Blanket
 from pyfecons.materials import Materials, Material
-from pyfecons.units import M_USD
+from pyfecons.units import M_USD, Meters3
 
 matplotlib.use('Agg')
+
+
+def compute_reactor_equipment_costs(reactor_type: ReactorType, blanket: Blanket, materials: Materials,
+                                    radial_build: RadialBuild) -> CAS220101:
+    IN = radial_build
+    OUT = CAS220101()
+    compute_inner_radii(reactor_type, IN, OUT)
+    compute_outer_radii(reactor_type, IN, OUT)
+    compute_volume(IN, OUT)
+
+    # must be cylindrical in all cases
+    OUT.gap2_vol = calc_volume_ring(IN.axis_t, IN.gap2_t + OUT.gap2_ir, OUT.gap2_ir)
+    # Updated bioshield volume
+    OUT.bioshield_vol = calc_volume_ring(IN.axis_t, IN.bioshield_t + OUT.bioshield_ir, OUT.bioshield_ir)
+
+    OUT.C22010101 = compute_first_wall_costs(blanket, materials, OUT)
+    OUT.C22010102 = compute_blanket_costs(blanket, materials, OUT)
+
+    # Total cost of blanket and first wall
+    OUT.C220101 = M_USD(OUT.C22010101 + OUT.C22010102)
+    return OUT
+
+
+def compute_volume(IN: RadialBuild, OUT: CAS220101) -> None:
+    OUT.axis_vol = 0.0
+    OUT.plasma_vol = Meters3(IN.elon * calc_volume_torus(IN.axis_t, OUT.plasma_ir, IN.plasma_t) - OUT.axis_vol)
+    OUT.vacuum_vol = Meters3(IN.elon * calc_volume_torus(IN.axis_t, OUT.vacuum_ir, IN.vacuum_t)
+                             - sum([OUT.plasma_vol, OUT.axis_vol]))
+    OUT.firstwall_vol = Meters3(IN.elon * calc_volume_torus(IN.axis_t, OUT.firstwall_ir, IN.firstwall_t)
+                                - sum([OUT.vacuum_vol, OUT.plasma_vol, OUT.axis_vol]))
+    OUT.blanket1_vol = Meters3(IN.elon * calc_volume_torus(IN.axis_t, OUT.blanket1_ir, IN.blanket1_t)
+                               - sum([OUT.firstwall_vol, OUT.vacuum_vol, OUT.plasma_vol, OUT.axis_vol]))
+    OUT.reflector_vol = Meters3(IN.elon * calc_volume_torus(IN.axis_t, OUT.reflector_ir, IN.reflector_t)
+                                - sum([OUT.blanket1_vol, OUT.firstwall_vol, OUT.vacuum_vol, OUT.plasma_vol,
+                                       OUT.axis_vol]))
+    OUT.ht_shield_vol = Meters3(IN.elon * calc_volume_torus(IN.axis_t, OUT.ht_shield_ir, IN.ht_shield_t)
+                                - sum([OUT.reflector_vol, OUT.blanket1_vol, OUT.firstwall_vol,
+                                       OUT.vacuum_vol, OUT.plasma_vol, OUT.axis_vol]))
+    OUT.structure_vol = Meters3(IN.elon * calc_volume_torus(IN.axis_t, OUT.structure_ir, IN.structure_t)
+                                - sum([OUT.ht_shield_vol, OUT.reflector_vol, OUT.blanket1_vol, OUT.firstwall_vol,
+                                       OUT.vacuum_vol, OUT.plasma_vol, OUT.axis_vol]))
+    OUT.gap1_vol = Meters3(IN.elon * calc_volume_torus(IN.axis_t, OUT.gap1_ir, IN.gap1_t)
+                           - sum([OUT.structure_vol, OUT.ht_shield_vol, OUT.reflector_vol, OUT.blanket1_vol,
+                                  OUT.firstwall_vol, OUT.vacuum_vol, OUT.plasma_vol, OUT.axis_vol]))
+    OUT.vessel_vol = Meters3(IN.elon * calc_volume_torus(IN.axis_t, OUT.vessel_ir, IN.vessel_t)
+                             - sum([OUT.gap1_vol, OUT.structure_vol, OUT.ht_shield_vol, OUT.reflector_vol,
+                                    OUT.blanket1_vol, OUT.firstwall_vol, OUT.vacuum_vol, OUT.plasma_vol, OUT.axis_vol]))
+    OUT.lt_shield_vol = Meters3(IN.elon * calc_volume_torus(IN.axis_t, OUT.lt_shield_ir, IN.lt_shield_t)
+                                - sum([OUT.vessel_vol, OUT.gap1_vol, OUT.structure_vol, OUT.ht_shield_vol,
+                                       OUT.reflector_vol, OUT.blanket1_vol, OUT.firstwall_vol, OUT.vacuum_vol,
+                                       OUT.plasma_vol, OUT.axis_vol]))
+    OUT.coil_vol = Meters3(calc_volume_torus(IN.axis_t, OUT.coil_ir, IN.coil_t) * 0.5)
 
 
 def compute_inner_radii(reactor_type: ReactorType, IN: RadialBuild, OUT: CAS220101):
